reduced nesting, changed naming, method return value
